requirements3.txt
asgiref==3.5.0
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
colorzero==1.1
cryptography==36.0.1
defusedxml==0.7.1
distro==1.5.0
Django==4.0.2
django-allauth==0.48.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-rest-framework==0.1.0
django-simpletwitchauth==1.0
djangorestframework==3.13.1
docopt==0.6.2
environ==1.0
gpiozero==1.6.2
idna==2.10
mysqlclient==2.1.0
oauthlib==3.2.0
pipreqs==0.4.11
pycparser==2.21
pyftdi==0.53.3
PyJWT==2.3.0
pyserial==3.5
python3-openid==3.2.0
pytz==2021.3
requests==2.25.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
ssh-import-id==5.10
urllib3==1.26.5
yarg==0.1.9